backend/requirements.txt
chromadb==1.4.1
fastapi==0.128.0
openai==2.15.0
playwright==1.55.0
psycopg2-binary==2.9.11
pyautogui==0.9.54
PyJWT==2.10.1
pytest==9.0.2
python-dotenv==1.2.1
python_bcrypt==0.3.2
pywhatkit==5.4
PyYAML==6.0.3
chromadb==1.4.1
fastapi==0.128.0
openai==2.15.0
playwright==1.55.0
psycopg2-binary==2.9.11
pyautogui==0.9.54
PyJWT==2.10.1
pytest==9.0.2
python-dotenv==1.2.1
python_bcrypt==0.3.2
pywhatkit==5.4
PyYAML==6.0.3
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
distro==1.9.0
filetype==1.2.0
google-auth==2.47.0
google-genai==1.58.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.7
langchain-google-genai==4.2.0
langsmith==0.6.3
orjson==3.11.5
packaging==25.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rsa==4.9.1
sniffio==1.3.1
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.13.0
websockets==15.0.1
zstandard==0.25.0
